Move variation diff construction into a separate file
